Upgrade to 7.5 and community edition image
The alpine version is no longer distributed, and there have been
some changes in the different plans that is available.

With this larger image more plugins are included, so we have dropped
those in the custom install as it is already bundled.

Note that we might have to remove the GitHub plugin, so I guess
we will have to deploy and see what happens...